Move GlobalStyles to components folder
