+++ components/private/GlobalStyles.tsx
